refactor: change from social-share to socialShare
