+++ app/utils/socialShare.ts
